* Code improved for 'm2s.c'.
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@470 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/libgpuarch/gpuarch.h
@@ -375,7 +375,7 @@ extern int gpu_pipeline_debug_category;
 struct gpu_t
 {
 	/* Current cycle */
-	uint64_t cycle;
+	long long cycle;
 
 	/* ND-Range running on it */
 	struct gpu_ndrange_t *ndrange;
